feat(rendering): add Lights plugin
