+++ engine/include/cubos/engine/render/lights/directional.hpp
@@ -1,6 +1,6 @@
 /// @file
-/// @brief Component @ref cubos::engine::PointLight.
-/// @ingroup renderer-plugin
+/// @brief Component @ref cubos::engine::DirectionalLight.
+/// @ingroup render-lights-plugin
 
 #pragma once
 
@@ -12,15 +12,16 @@
 
 namespace cubos::engine
 {
-    /// @brief Component which makes an entity behave like a point light.
-    /// @note Should be used with @ref LocalToWorld.
-    /// @ingroup renderer-plugin
-    struct CUBOS_ENGINE_API PointLight
+    /// @brief Component which makes an entity behave like a directional light.
+    ///
+    /// The light points towards the local forward vector of the entity.
+    ///
+    /// @ingroup render-lights-plugin
+    struct CUBOS_ENGINE_API DirectionalLight
     {
         CUBOS_REFLECT;
 
         glm::vec3 color = {1.0F, 1.0F, 1.0F}; ///< Color of the light.
         float intensity = 1.0F;               ///< Intensity of the light.
-        float range = 5.0F;                   ///< Range of the light.
     };
 } // namespace cubos::engine
